Update camera.md to remove reference to rpi-update
I believe the Camera works fine with the latest available stable firmware, and there's no reason to suggest 'rpi-update' as a debugging step since it could very easily have unexpected repercussions, particularly with changes like Device Tree. At the time of writing, this step would absolutely cause more problems than it would solve, and that's likely to be the same in future.

I don't know how robust dist-upgrade is as a way to get the latest firmware, or whether it's served as part of "apt-get upgrade", so perhaps this section should be ditched altogether.